requirements.txt
גרסת פייתון- 3.12

חבילות שהתקנו:
matplotlib-3.10.0
numpy-2.2.2
pyshark-0.6